move bulkremap button & add set to none button
