+++ .eslintrc.js
@@ -42,6 +42,10 @@ module.exports = {
       allowFunctions: false,
       allowBind: false,
     }],
-    "@typescript-eslint/no-shadow": 1
+    "@typescript-eslint/no-shadow": 1,
+    "linebreak-style": [
+      "error",
+      "windows"
+    ],
   }
 };
